pyproject.toml
[project]
name = "internship-intel"
version = "0.1.0"
description = "AI/ML internship discovery and application pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.6",
    "alembic>=1.14.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sentence-transformers>=3.3.0",
    "groq>=0.12.0",
    "google-generativeai>=0.8.0",
    "tenacity>=9.0.0",
    "httpx>=0.28.0",
    "selectolax>=0.3.26",
    "apscheduler>=3.10.0",
    "cryptography>=43.0.0",
    "passlib[argon2]>=1.7.4",
    "itsdangerous>=2.2.0",
    "structlog>=24.4.0",
    "numpy>=1.26.0",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "hypothesis>=6.122.0",
    "polyfactory>=2.18.0",
    "respx>=0.21.0",
    "vulture>=2.13",
    "ruff>=0.8.0",
    "anyio>=4.6.0",
    "freezegun>=1.5.5",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
